feat: HELMSMAN_CUSTOM_HOOKS_DIR env. variable overrides custom hooks location
By default hooks are read from the `custom-hooks` directory. You can
override this by setting the `HELMSMAN_CUSTOM_HOOKS_DIR` environment
variable.

Alternatively, you can set the `HELMSMAN_CUSTOM_HOOKS_DIR` environment
variable to point to a different directory. This is useful when you want
to use a private repository with this project as a submodule.