fix: extend maintenance-worker process lifetime to 1 hour and put it in a loop
Previously, the process was limited to a 5-minute lifetime before its pod
was restarted. This commit modifies the configuration so that the pod
remains active indefinitely, and only the pimcore's messenger:consume
process is restarted every hour.

This change alleviates the unnecessary burden on the Kubernetes cluster
and reduces logging overhead.